Rename Timestampable* traits to Timestamped
